requirements.txt
opencv-python==4.1.1.26
lxml
tqdm

absl-py
easydict
matplotlib
pillow
pytesseract

imutils
numpy
tensorflow-gpu==2.3.0rc0


dnspython
Flask
Flask-PyMongo